pyproject.toml
[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[project]
name = "doubletapp-django"
version = "0.1.0"
description = ""
authors = [
    {name = "frukun",email = "peromata@yandex.ru"}
]
readme = "README.md"
requires-python = "3.13.3"
dependencies = [
    "django (==5.1.7)",
    "django-environ (==0.12.0)",
    "phonenumbers (==9.0.0)",
    "python-telegram-bot[webhooks] (==22.0)",
    "psycopg[binary] (==3.2.5)",
    "gunicorn (==23.0.0)",
    "django-ninja (==1.4.1)",
    "django-ninja-extra (==0.30.0)",
    "django-ninja-jwt (==5.3.7)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.1.2"
isort = "^6.0.1"
debugpy = "^1.8.13"
pytest = "^8.3.5"
pytest-django = "^4.10.0"
pytest-asyncio = "^0.26.0"
freezegun = "^1.5.1"
pytest-xdist = "^3.6.1"

[tool.pytest.ini_options]
pythonpath = "src"
DJANGO_SETTINGS_MODULE = "config.settings"
testpaths = [
    "tests",
]
addopts = "--reuse-db --create-db --nomigrations"
markers = [
    "integration: integration test",
    "smoke: smoke test",
    "unit: unit test"
]
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
